+++ routes/security/functional_areas.rb
@@ -37,6 +37,7 @@ class LabelDesigner < Roda
         end
         r.delete do    # DELETE
           return_json_response
+          raise Crossbeams::AuthorizationError unless authorised?('menu', 'delete')
           res = interactor.delete_functional_area(id)
           flash[:notice] = res.message
           redirect_to_last_grid(r)
@@ -54,24 +55,12 @@ class LabelDesigner < Roda
         res = interactor.create_functional_area(params[:functional_area])
         if res.success
           flash[:notice] = res.message
-          if fetch?(r)
-            redirect_via_json_to_last_grid
-          else
-            redirect_to_last_grid(r)
-          end
-        elsif fetch?(r)
-          content = show_partial do
-            Security::FunctionalAreas::FunctionalArea::New.call(form_values: params[:functional_area],
-                                                                form_errors: res.errors,
-                                                                remote: true)
-          end
-          update_dialog_content(content: content, error: res.message)
+          redirect_to_last_grid(r)
         else
-          flash[:error] = res.message
-          show_page do
+          re_show_form(r, res, url: '/security/functional_areas/functional_areas/new') do
             Security::FunctionalAreas::FunctionalArea::New.call(form_values: params[:functional_area],
                                                                 form_errors: res.errors,
-                                                                remote: false)
+                                                                remote: fetch?(r))
           end
         end
       end
@@ -150,26 +139,13 @@ class LabelDesigner < Roda
         res = interactor.create_program(params[:program], self.class.name)
         if res.success
           flash[:notice] = res.message
-          if fetch?(r)
-            redirect_via_json_to_last_grid
-          else
-            redirect_to_last_grid(r)
-          end
-        elsif fetch?(r)
-          content = show_partial do # params[:program][:functional_area_id]
-            Security::FunctionalAreas::Program::New.call(res.functional_area_id,
-                                                         form_values: params[:program],
-                                                         form_errors: res.errors,
-                                                         remote: true)
-          end
-          update_dialog_content(content: content, error: res.message)
+          redirect_to_last_grid(r)
         else
-          flash[:error] = res.message
-          show_page do
+          re_show_form(r, res, url: "/security/functional_areas/programs/#{res.functional_area_id}/new") do
             Security::FunctionalAreas::Program::New.call(res.functional_area_id,
                                                          form_values: params[:program],
                                                          form_errors: res.errors,
-                                                         remote: false)
+                                                         remote: fetch?(r))
           end
         end
       end
@@ -234,26 +210,13 @@ class LabelDesigner < Roda
         res = interactor.create_program_function(params[:program_function])
         if res.success
           flash[:notice] = res.message
-          if fetch?(r)
-            redirect_via_json_to_last_grid
-          else
-            redirect_to_last_grid(r)
-          end
-        elsif fetch?(r)
-          content = show_partial do
-            Security::FunctionalAreas::ProgramFunction::New.call(params[:program_function][:program_id],
-                                                                 form_values: params[:program_function],
-                                                                 form_errors: res.errors,
-                                                                 remote: true)
-          end
-          update_dialog_content(content: content, error: res.message)
+          redirect_to_last_grid(r)
         else
-          flash[:error] = res.message
-          show_page do
+          re_show_form(r, res, url: "/security/functional_areas/program_functions/#{params[:program_function][:program_id]}/new") do
             Security::FunctionalAreas::ProgramFunction::New.call(params[:program_function][:program_id],
                                                                  form_values: params[:program_function],
                                                                  form_errors: res.errors,
-                                                                 remote: false)
+                                                                 remote: fetch?(r))
           end
         end
       end
@@ -308,6 +271,7 @@ class LabelDesigner < Roda
         end
         r.delete do    # DELETE
           return_json_response
+          raise Crossbeams::AuthorizationError unless authorised?('menu', 'delete')
           res = interactor.delete_security_group(id)
           delete_grid_row(id, notice: res.message)
         end
@@ -323,24 +287,12 @@ class LabelDesigner < Roda
         res = interactor.create_security_group(params[:security_group])
         if res.success
           flash[:notice] = res.message
-          if fetch?(r)
-            redirect_via_json_to_last_grid
-          else
-            redirect_to_last_grid(r)
-          end
-        elsif fetch?(r)
-          content = show_partial do
-            Security::FunctionalAreas::SecurityGroup::New.call(form_values: params[:security_group],
-                                                               form_errors: res.errors,
-                                                               remote: true)
-          end
-          update_dialog_content(content: content, error: res.message)
+          redirect_to_last_grid(r)
         else
-          flash[:error] = res.message
-          show_page do
+          re_show_form(r, res, url: '/security/functional_areas/security_groups/new') do
             Security::FunctionalAreas::SecurityGroup::New.call(form_values: params[:security_group],
                                                                form_errors: res.errors,
-                                                               remote: false)
+                                                               remote: fetch?(r))
           end
         end
       end
@@ -396,24 +348,12 @@ class LabelDesigner < Roda
         res = interactor.create_security_permission(params[:security_permission])
         if res.success
           flash[:notice] = res.message
-          if fetch?(r)
-            redirect_via_json_to_last_grid
-          else
-            redirect_to_last_grid(r)
-          end
-        elsif fetch?(r)
-          content = show_partial do
-            Security::FunctionalAreas::SecurityPermission::New.call(form_values: params[:security_permission],
-                                                                    form_errors: res.errors,
-                                                                    remote: true)
-          end
-          update_dialog_content(content: content, error: res.message)
+          redirect_to_last_grid(r)
         else
-          flash[:error] = res.message
-          show_page do
+          re_show_form(r, res, url: '/security/functional_areas/security_permissions/new') do
             Security::FunctionalAreas::SecurityPermission::New.call(form_values: params[:security_permission],
                                                                     form_errors: res.errors,
-                                                                    remote: false)
+                                                                    remote: fetch?(r))
           end
         end
       end
